cuda use fp16 for all

+++ include/conv.h
@@ -35,8 +35,8 @@ void nn_upsample_nhwc(const float* input,
 #ifdef __CUDACC__
 #include <cuda_fp16.h>
 
-__global__ void conv_relu_nhwc_oihw_cuda(const float* input,
-                        float* output,
+__global__ void conv_relu_nhwc_oihw_cuda(const half* input,
+                        half* output,
                         size_t in_h,
                         size_t in_w,
                         size_t filter_h,
@@ -46,20 +46,20 @@ __global__ void conv_relu_nhwc_oihw_cuda(const float* input,
                         const __half* weights,
                         const __half* bias);
 
-__global__ void max_pool_nhwc_cuda(const float* input,
-                   float* output,
+__global__ void max_pool_nhwc_cuda(const half* input,
+                   half* output,
                    size_t in_h,
                    size_t in_w,
                    size_t in_c);
 
-__global__ void avg_pool_nhwc_cuda(const float* input,
-                   float* output,
+__global__ void avg_pool_nhwc_cuda(const half* input,
+                   half* output,
                    size_t in_h,
                    size_t in_w,
                    size_t in_c);
 
-__global__ void nn_upsample_nhwc_cuda(const float* input,
-                      float* output,
+__global__ void nn_upsample_nhwc_cuda(const half* input,
+                      half* output,
                       size_t in_h,
                       size_t in_w,
                       size_t in_c);
